refactor(managers.py): move funcs into biz or other layers
