+++ src/dashboard/apigateway/apigateway/apis/web/sdk/views.py
@@ -19,6 +19,7 @@
 from typing import Any, Dict, cast
 
 from django.db import transaction
+from django.db.models import Q
 from django.shortcuts import get_object_or_404
 from django.utils.decorators import method_decorator
 from django.utils.translation import gettext as _
@@ -61,7 +62,7 @@ class GatewaySDKListCreateApi(generics.ListCreateAPIView):
         slz = serializers.GatewaySDKQueryInputSLZ(data=request.query_params, context={"request": request})
         slz.is_valid(raise_exception=True)
 
-        queryset = GatewaySDK.objects.filter_sdk(
+        queryset = self._filter_sdk(
             gateway=self.request.gateway,
             language=slz.validated_data.get("language"),
             version_number=slz.validated_data.get("version_number"),
@@ -76,6 +77,42 @@ class GatewaySDKListCreateApi(generics.ListCreateAPIView):
         sdks = [SDKFactory.create(model=i) for i in page]
         slz = self.get_serializer(sdks, many=True)
         return self.get_paginated_response(slz.data)
+
+    def _filter_sdk(
+        self,
+        gateway,
+        language=None,
+        order_by=None,
+        version_number="",
+        resource_version_id=None,
+        fuzzy=False,
+        keyword=None,
+    ):
+        queryset = GatewaySDK.objects.filter(gateway=gateway)
+
+        if keyword:
+            queryset = queryset.filter(
+                Q(language__icontains=keyword)
+                | Q(version_number__contains=keyword)
+                | Q(resource_version__version__contains=keyword)
+            )
+
+        if language:
+            queryset = queryset.filter(language=language)
+
+        if version_number:
+            if fuzzy:
+                queryset = queryset.filter(version_number__contains=version_number)
+            else:
+                queryset = queryset.filter(version_number=version_number)
+
+        if resource_version_id is not None:
+            queryset = queryset.filter(resource_version_id=resource_version_id)
+
+        if order_by:
+            queryset = queryset.order_by(order_by)
+
+        return queryset
 
     @transaction.atomic
     def create(self, request, gateway_id):
