+++ src/dashboard/apigateway/apigateway/apps/support/managers.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # TencentBlueKing is pleased to support the open source community by making
-# 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
+# 蓝鲸智云 - API 网关 (BlueKing - APIGateway) available.
 # Copyright (C) 2017 THL A29 Limited, a Tencent company. All rights reserved.
 # Licensed under the MIT License (the "License"); you may not use this file except
 # in compliance with the License. You may obtain a copy of the License at
@@ -16,11 +16,9 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-from collections import defaultdict
-from typing import Any, Dict, List, Optional
+from typing import Dict
 
 from django.db import models
-from django.db.models import Q
 
 from apigateway.apps.support.constants import DocLanguageEnum
 from apigateway.core.constants import GatewayStatusEnum
@@ -29,47 +27,11 @@ RELEASED_RESOURCE_DOC_CREATE_BATCH_SIZE = 50
 
 
 class ResourceDocVersionManager(models.Manager):
-    def make_version(self, gateway_id):
-        from apigateway.apps.support.models import ResourceDoc
-
-        docs = ResourceDoc.objects.filter(gateway_id=gateway_id).all()
-        return [d.snapshot(as_dict=True) for d in docs]
-
     def get_by_resource_version_id(self, gateway_id: int, resource_version_id: int):
         return self.filter(gateway_id=gateway_id, resource_version_id=resource_version_id).first()
 
     def get_latest_version(self, gateway_id):
         return self.filter(gateway_id=gateway_id).order_by("-id").first()
-
-    def get_doc_data_by_rv_or_new(self, gateway_id: int, resource_version_id: Optional[int]) -> List[Any]:
-        """获取版本中文档内容"""
-        if resource_version_id:
-            try:
-                return self.get(gateway_id=gateway_id, resource_version_id=resource_version_id).data
-            except self.model.DoesNotExist:
-                return []
-
-        return self.make_version(gateway_id)
-
-    def get_doc_updated_time(self, gateway_id: int, resource_version_id: Optional[int]):
-        """获取文档更新时间
-
-        @return:
-        {
-            1: {
-                "zh": "1970-01-01 12:30:50 +8000",
-                "en": "1970-01-01 12:30:50 +8000"
-            }
-        }
-        """
-        doc_data = self.get_doc_data_by_rv_or_new(gateway_id, resource_version_id)
-
-        result: Dict[int, Dict[str, Any]] = defaultdict(dict)
-        for doc in doc_data:
-            language = doc.get("language", DocLanguageEnum.ZH.value)
-            result[doc["resource_id"]][language] = doc["updated_time"]
-
-        return result
 
 
 class ReleasedResourceDocManager(models.Manager):
@@ -97,61 +59,17 @@ class ReleasedResourceDocManager(models.Manager):
             )
             for doc in resource_doc_version.data
         ]
-        # 异步同时(多个stage同时发布同一版本)更新会存在一些冲突问题
+        # 异步同时 (多个 stage 同时发布同一版本) 更新会存在一些冲突问题
         self.bulk_create(
             resource_doc_to_add, batch_size=RELEASED_RESOURCE_DOC_CREATE_BATCH_SIZE, ignore_conflicts=True
         )
-
-    def clear_unreleased_resource_doc(self, gateway_id: int) -> None:
-        """清理未发布的资源文档，如已发布版本被新版本替代的情况"""
-        from apigateway.core.models import Release
-
-        resource_version_ids = Release.objects.get_released_resource_version_ids(gateway_id)
-        self.filter(gateway_id=gateway_id).exclude(resource_version_id__in=resource_version_ids).delete()
 
     def get_doc_updated_time(self, gateway_id: int, resource_version_id: int, resource_id: int) -> Dict[str, str]:
         qs = self.filter(gateway_id=gateway_id, resource_version_id=resource_version_id, resource_id=resource_id)
         return {doc.language: doc.data["updated_time"] for doc in qs}
 
 
-class APISDKManager(models.Manager):
-    # FIXME: move to views.py
-    def filter_sdk(
-        self,
-        gateway,
-        language=None,
-        order_by=None,
-        version_number="",
-        resource_version_id=None,
-        fuzzy=False,
-        keyword=None,
-    ):
-        queryset = self.filter(gateway=gateway)
-
-        if keyword:
-            queryset = queryset.filter(
-                Q(language__icontains=keyword)
-                | Q(version_number__contains=keyword)
-                | Q(resource_version__version__contains=keyword)
-            )
-
-        if language:
-            queryset = queryset.filter(language=language)
-
-        if version_number:
-            if fuzzy:
-                queryset = queryset.filter(version_number__contains=version_number)
-            else:
-                queryset = queryset.filter(version_number=version_number)
-
-        if resource_version_id is not None:
-            queryset = queryset.filter(resource_version_id=resource_version_id)
-
-        if order_by:
-            queryset = queryset.order_by(order_by)
-
-        return queryset
-
+class GatewaySDKManager(models.Manager):
     def get_latest_sdk(self, gateway_id, language):
         return self.filter(gateway_id=gateway_id, language=language).order_by("-id").first()
 
@@ -161,12 +79,12 @@ class APISDKManager(models.Manager):
     def should_be_set_to_public_latest(
         self, gateway_id: int, resource_version_id: int, is_uploaded_to_pypi: bool
     ) -> bool:
-        """是否应该将此SDK设置为网关最新公开的SDK"""
+        """是否应该将此 SDK 设置为网关最新公开的 SDK"""
         if not is_uploaded_to_pypi:
             return False
 
         public_latest_sdk = self.filter(is_recommended=True, gateway_id=gateway_id).first()
-        # 最新 SDK 的版本比当前版本更新，则当前SDK非最新SDK
+        # 最新 SDK 的版本比当前版本更新，则当前 SDK 非最新 SDK
         if (
             public_latest_sdk
             and public_latest_sdk.resource_version_id
@@ -178,7 +96,7 @@ class APISDKManager(models.Manager):
 
     def filter_recommended_sdks(self, language, gateway_id=None):
         """
-        获取公开网关的最新公开SDK
+        获取公开网关的最新公开 SDK
         """
         queryset = self.filter(
             is_recommended=True,
